+++ encodings/sparse/src/lib.rs
@@ -182,6 +182,8 @@ impl VTable for SparseVTable {
         }
 
         // SAFETY:
+        // patches slice will ensure that dtype of patches is unchanged and the indices and
+        // values match
         Ok(Some(
             unsafe { SparseArray::new_unchecked(new_patches, array.fill_scalar().clone()) }
                 .into_array(),
