+++ vortex-array/src/vtable/mod.rs
@@ -190,8 +190,7 @@ pub trait VTable: 'static + Sized + Send + Sync + Debug {
 
     /// Perform a constant-time slice of the array.
     ///
-    /// If an encoding cannot perform this slice in constant time, it should instead
-    /// wrap itself in the `SliceArray`.
+    /// If an encoding cannot perform this slice in constant time, it should instead return Ok(None).
     ///
     /// This function returns [`ArrayRef`] since some encodings can return a simpler array for
     /// some slices, for example a [`crate::arrays::ChunkedArray`] may slice into a single chunk.
